Move screenshots below links to introductory essays
Apparently many non-users that take part in discussions about Magit
cannot be bothered to look at anything beyond the first block of text.
We try to deal with that by making these very important links part of
the initial block of text.

I think these two texts of mine are actually rather good and it annoys
me a great deal that there's always someone will "I read the homepage
but don't understand what this is all about", when all they have to do
is click on one of these links and keep reading.